// src/app/modules/hr/models/employee.ts
export class Employee {

  id: number;
  photo: string;
  code: string;
  name: string;
  fatherName: string;
  dob: Date;
  doj: Date;
  nid: string;
  gender: string;
  email: string;
  mobileNo: string;
  active: boolean;
  userId: number;
  deptId: number;
  desgId: number;
  supervisorId: number;

}